receipt_recognition/requirements.txt
streamlit
transformers
torch
langchain
safetensors
Pillow